[chore] move student plan service to usecase
